admin/requirements.txt
flask==3.0.2
werkzeug==3.0.1
flask-session==0.5.0
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==4.0.0
GitPython==3.1.42 